Test: Name Update and L5.8 changes

+++ database/migrations/2016_07_29_171118_create_chatter_post_table.php
@@ -8,9 +8,9 @@ class CreateChatterPostTable extends Migration
     public function up()
     {
         Schema::create('chatter_post', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('chatter_discussion_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('chatter_discussion_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->text('body');
             $table->timestamps();
         });
